+++ paddle_billing/Entities/Subscriptions/SubscriptionConsentRequirementStatus.py
@@ -1,0 +1,7 @@
+from paddle_billing.PaddleStrEnum import PaddleStrEnum, PaddleStrEnumMeta
+
+
+class SubscriptionConsentRequirementStatus(PaddleStrEnum, metaclass=PaddleStrEnumMeta):
+    Pending: "SubscriptionConsentRequirementStatus" = "pending"
+    Granted: "SubscriptionConsentRequirementStatus" = "granted"
+    Voided: "SubscriptionConsentRequirementStatus" = "voided"