+++ paddle_billing/Entities/Subscriptions/SubscriptionDiscountType.py
@@ -1,0 +1,6 @@
+from paddle_billing.PaddleStrEnum import PaddleStrEnum, PaddleStrEnumMeta
+
+
+class SubscriptionDiscountType(PaddleStrEnum, metaclass=PaddleStrEnumMeta):
+    Recurring: "SubscriptionDiscountType" = "recurring"
+    OneOff: "SubscriptionDiscountType" = "one-off"