+++ CHANGELOG.md
@@ -20,6 +20,9 @@ Check our main [developer changelog](https://developer.paddle.com/?utm_source=dx
 - Added `CLP` and `PEN` currency code support. See [related changelog](https://developer.paddle.com/changelog/2026/clp-pen-currencies?utm_source=dx&utm_medium=paddle-python-sdk)
 - Added `rotatable` to api_key events. See [related changelog](https://developer.paddle.com/changelog/2026/aws-secrets-manager-rotation?utm_source=dx&utm_medium=paddle-python-sdk)
 - Added `balance_movement_date` and `balance_movement_type` filters for payout reconciliation reports. See [related changelog](https://developer.paddle.com/changelog/2026/payout-reconciliation-report-filters?utm_source=dx&utm_medium=paddle-python-sdk)
+- Added `type` to subscription `discount`
+- Added `consent_requirements` to subscriptions
+- Added `next_billed_at` filter to list subscriptions operation
 
 ### Changed